Simplify import path
