+++ tests/_hashing/file_hashing_test.py
@@ -16,7 +16,7 @@ import pathlib
 
 import pytest
 
-from model_signing._hashing import file
+from model_signing._hashing import file_hashing
 from model_signing._hashing import memory
 
 
@@ -68,22 +68,24 @@ def expected_content_digest():
 class TestSimpleFileHasher:
     def test_fails_with_negative_chunk_size(self):
         with pytest.raises(ValueError, match="Chunk size must be non-negative"):
-            file.SimpleFileHasher(_UNUSED_PATH, memory.SHA256(), chunk_size=-2)
+            file_hashing.SimpleFileHasher(
+                _UNUSED_PATH, memory.SHA256(), chunk_size=-2
+            )
 
     def test_hash_of_known_file(self, sample_file, expected_digest):
-        hasher = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest = hasher.compute()
         assert digest.digest_hex == expected_digest
 
     def test_hash_of_known_file_no_chunk(self, sample_file, expected_digest):
-        hasher = file.SimpleFileHasher(
+        hasher = file_hashing.SimpleFileHasher(
             sample_file, memory.SHA256(), chunk_size=0
         )
         digest = hasher.compute()
         assert digest.digest_hex == expected_digest
 
     def test_hash_of_known_file_small_chunk(self, sample_file, expected_digest):
-        hasher = file.SimpleFileHasher(
+        hasher = file_hashing.SimpleFileHasher(
             sample_file, memory.SHA256(), chunk_size=2
         )
         digest = hasher.compute()
@@ -91,34 +93,34 @@ class TestSimpleFileHasher:
 
     def test_hash_of_known_file_large_chunk(self, sample_file, expected_digest):
         size = 2 * len(_FULL_CONTENT)
-        hasher = file.SimpleFileHasher(
+        hasher = file_hashing.SimpleFileHasher(
             sample_file, memory.SHA256(), chunk_size=size
         )
         digest = hasher.compute()
         assert digest.digest_hex == expected_digest
 
     def test_hash_file_twice(self, sample_file):
-        hasher1 = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher1 = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest1 = hasher1.compute()
-        hasher2 = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher2 = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest2 = hasher2.compute()
         assert digest1.digest_value == digest2.digest_value
 
     def test_hash_file_twice_same_hasher(self, sample_file):
-        hasher = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest1 = hasher.compute()
         digest2 = hasher.compute()
         assert digest1.digest_value == digest2.digest_value
 
     def test_hash_file_twice_same_hasher_reset_file(self, sample_file):
-        hasher = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest1 = hasher.compute()
         hasher.set_file(sample_file)
         digest2 = hasher.compute()
         assert digest1.digest_value == digest2.digest_value
 
     def test_set_file(self, sample_file, sample_file_content_only):
-        hasher = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest1 = hasher.compute()
         hasher.set_file(sample_file_content_only)
         _ = hasher.compute()
@@ -127,11 +129,11 @@ class TestSimpleFileHasher:
         assert digest1.digest_value == digest2.digest_value
 
     def test_default_digest_name(self):
-        hasher = file.SimpleFileHasher(_UNUSED_PATH, memory.SHA256())
+        hasher = file_hashing.SimpleFileHasher(_UNUSED_PATH, memory.SHA256())
         assert hasher.digest_name == "sha256"
 
     def test_override_digest_name(self):
-        hasher = file.SimpleFileHasher(
+        hasher = file_hashing.SimpleFileHasher(
             _UNUSED_PATH,
             memory.SHA256(),
             chunk_size=10,
@@ -140,13 +142,13 @@ class TestSimpleFileHasher:
         assert hasher.digest_name == "test-hash"
 
     def test_digest_algorithm_is_digest_name(self, sample_file):
-        hasher = file.SimpleFileHasher(sample_file, memory.SHA256())
+        hasher = file_hashing.SimpleFileHasher(sample_file, memory.SHA256())
         digest = hasher.compute()
         assert digest.algorithm == hasher.digest_name
 
     def test_digest_size(self):
         memory_hasher = memory.SHA256()
-        hasher = file.SimpleFileHasher(sample_file, memory_hasher)
+        hasher = file_hashing.SimpleFileHasher(sample_file, memory_hasher)
         assert hasher.digest_size == memory_hasher.digest_size
 
 
@@ -155,7 +157,7 @@ class TestShardedFileHasher:
         with pytest.raises(
             ValueError, match="Shard size must be strictly positive"
         ):
-            file.ShardedFileHasher(
+            file_hashing.ShardedFileHasher(
                 _UNUSED_PATH, memory.SHA256(), shard_size=-2, start=0, end=42
             )
 
@@ -163,12 +165,12 @@ class TestShardedFileHasher:
         with pytest.raises(
             ValueError, match="File start offset must be non-negative"
         ):
-            file.ShardedFileHasher(
+            file_hashing.ShardedFileHasher(
                 _UNUSED_PATH, memory.SHA256(), start=-2, end=42
             )
 
     def test_set_fails_with_negative_start(self):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             _UNUSED_PATH, memory.SHA256(), start=0, end=42
         )
         with pytest.raises(
@@ -183,12 +185,12 @@ class TestShardedFileHasher:
                 "File end offset must be stricly higher that file start offset"
             ),
         ):
-            file.ShardedFileHasher(
+            file_hashing.ShardedFileHasher(
                 _UNUSED_PATH, memory.SHA256(), start=42, end=2
             )
 
     def test_set_fails_with_end_lower_than_start(self):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             _UNUSED_PATH, memory.SHA256(), start=0, end=42
         )
         with pytest.raises(
@@ -206,12 +208,12 @@ class TestShardedFileHasher:
                 "File end offset must be stricly higher that file start offset"
             ),
         ):
-            file.ShardedFileHasher(
+            file_hashing.ShardedFileHasher(
                 _UNUSED_PATH, memory.SHA256(), start=42, end=42
             )
 
     def test_set_fails_with_zero_read_span(self):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             _UNUSED_PATH, memory.SHA256(), start=0, end=42
         )
         with pytest.raises(
@@ -226,12 +228,12 @@ class TestShardedFileHasher:
         with pytest.raises(
             ValueError, match="Must not read more than shard_size=2"
         ):
-            file.ShardedFileHasher(
+            file_hashing.ShardedFileHasher(
                 _UNUSED_PATH, memory.SHA256(), start=0, end=42, shard_size=2
             )
 
     def test_set_fails_with_read_span_too_large(self):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             _UNUSED_PATH, memory.SHA256(), start=0, end=2, shard_size=2
         )
         with pytest.raises(
@@ -242,10 +244,10 @@ class TestShardedFileHasher:
     def test_hash_of_known_file(
         self, sample_file, expected_header_digest, expected_content_digest
     ):
-        hasher1 = file.ShardedFileHasher(
+        hasher1 = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=0, end=_SHARD_SIZE
         )
-        hasher2 = file.ShardedFileHasher(
+        hasher2 = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=_SHARD_SIZE, end=2 * _SHARD_SIZE
         )
 
@@ -258,7 +260,7 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_using_set_shard(
         self, sample_file, expected_header_digest, expected_content_digest
     ):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=0, end=_SHARD_SIZE
         )
 
@@ -272,7 +274,7 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_end_overflow(
         self, sample_file, expected_digest
     ):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=0, end=3 * _SHARD_SIZE
         )
         digest = hasher.compute()
@@ -281,7 +283,7 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_set_end_overflow(
         self, sample_file, expected_digest
     ):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=0, end=_SHARD_SIZE
         )
         hasher.set_shard(start=0, end=5 * _SHARD_SIZE)
@@ -291,10 +293,10 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_no_chunk(
         self, sample_file, expected_header_digest, expected_content_digest
     ):
-        hasher1 = file.ShardedFileHasher(
+        hasher1 = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=0, end=_SHARD_SIZE, chunk_size=0
         )
-        hasher2 = file.ShardedFileHasher(
+        hasher2 = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=_SHARD_SIZE,
@@ -311,10 +313,10 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_small_chunk(
         self, sample_file, expected_header_digest, expected_content_digest
     ):
-        hasher1 = file.ShardedFileHasher(
+        hasher1 = file_hashing.ShardedFileHasher(
             sample_file, memory.SHA256(), start=0, end=_SHARD_SIZE, chunk_size=1
         )
-        hasher2 = file.ShardedFileHasher(
+        hasher2 = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=_SHARD_SIZE,
@@ -331,14 +333,14 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_large_chunk(
         self, sample_file, expected_header_digest, expected_content_digest
     ):
-        hasher1 = file.ShardedFileHasher(
+        hasher1 = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=0,
             end=_SHARD_SIZE,
             chunk_size=2 * len(_FULL_CONTENT),
         )
-        hasher2 = file.ShardedFileHasher(
+        hasher2 = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=_SHARD_SIZE,
@@ -355,14 +357,14 @@ class TestShardedFileHasher:
     def test_hash_of_known_file_large_shard(
         self, sample_file, expected_header_digest, expected_content_digest
     ):
-        hasher1 = file.ShardedFileHasher(
+        hasher1 = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=0,
             end=_SHARD_SIZE,
             shard_size=2 * len(_FULL_CONTENT),
         )
-        hasher2 = file.ShardedFileHasher(
+        hasher2 = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=_SHARD_SIZE,
@@ -377,13 +379,13 @@ class TestShardedFileHasher:
         assert digest2.digest_hex == expected_content_digest
 
     def test_default_digest_name(self):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             _UNUSED_PATH, memory.SHA256(), start=0, end=2, shard_size=10
         )
         assert hasher.digest_name == "sha256-sharded-10"
 
     def test_override_digest_name(self):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             _UNUSED_PATH,
             memory.SHA256(),
             start=0,
@@ -394,7 +396,7 @@ class TestShardedFileHasher:
         assert hasher.digest_name == "test-hash"
 
     def test_digest_algorithm_is_digest_name(self, sample_file):
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             sample_file,
             memory.SHA256(),
             start=0,
@@ -407,7 +409,7 @@ class TestShardedFileHasher:
 
     def test_digest_size(self):
         memory_hasher = memory.SHA256()
-        hasher = file.ShardedFileHasher(
+        hasher = file_hashing.ShardedFileHasher(
             sample_file, memory_hasher, start=0, end=2
         )
         assert hasher.digest_size == memory_hasher.digest_size
