+++ tests/_serialization/testdata/file/TestSerializer/empty_model_file
