+++ tests/_serialization/testdata/file/TestSerializer/model_folder_with_empty_file
