+++ tests/_serialization/testdata/file/TestSerializer/symlink_model_folder
