+++ tests/_serialization/testdata/file_shard/TestSerializer/deep_model_folder_small_shards
