+++ tests/_serialization/testdata/file_shard/TestSerializer/sample_model_file
