+++ tests/_serialization/testdata/file_shard/TestSerializer/sample_model_file_small_shards
