+++ tests/_serialization/testdata/file_shard/TestSerializer/symlink_model_folder
