+++ tests/signing/in_toto_signature_test.py
@@ -14,10 +14,10 @@
 
 import pathlib
 
-from model_signing._hashing import file
+from model_signing._hashing import file_hashing
 from model_signing._hashing import memory
-from model_signing._serialization import serialize_by_file
-from model_signing._serialization import serialize_by_file_shard
+from model_signing._serialization import file
+from model_signing._serialization import file_shard
 from model_signing.signature import fake
 from model_signing.signing import in_toto
 from model_signing.signing import in_toto_signature
@@ -26,18 +26,18 @@ from model_signing.signing import in_toto_signature
 class TestIntotoSignature:
     def _shard_hasher_factory(
         self, path: pathlib.Path, start: int, end: int
-    ) -> file.ShardedFileHasher:
-        return file.ShardedFileHasher(
+    ) -> file_hashing.ShardedFileHasher:
+        return file_hashing.ShardedFileHasher(
             path, memory.SHA256(), start=start, end=end
         )
 
-    def _hasher_factory(self, path: pathlib.Path) -> file.FileHasher:
-        return file.SimpleFileHasher(path, memory.SHA256())
+    def _hasher_factory(self, path: pathlib.Path) -> file_hashing.FileHasher:
+        return file_hashing.SimpleFileHasher(path, memory.SHA256())
 
     def test_sign_and_verify_sharded_manifest(self, sample_model_folder):
         signer = in_toto_signature.IntotoSigner(fake.FakeSigner())
         verifier = in_toto_signature.IntotoVerifier(fake.FakeVerifier())
-        shard_serializer = serialize_by_file_shard.Serializer(
+        shard_serializer = file_shard.Serializer(
             self._shard_hasher_factory, allow_symlinks=True
         )
         shard_manifest = shard_serializer.serialize(sample_model_folder)
@@ -53,7 +53,7 @@ class TestIntotoSignature:
     def test_sign_and_verify_digest_sharded_manifest(self, sample_model_folder):
         signer = in_toto_signature.IntotoSigner(fake.FakeSigner())
         verifier = in_toto_signature.IntotoVerifier(fake.FakeVerifier())
-        shard_serializer = serialize_by_file_shard.Serializer(
+        shard_serializer = file_shard.Serializer(
             self._shard_hasher_factory, allow_symlinks=True
         )
         shard_manifest = shard_serializer.serialize(sample_model_folder)
@@ -71,7 +71,7 @@ class TestIntotoSignature:
     ):
         signer = in_toto_signature.IntotoSigner(fake.FakeSigner())
         verifier = in_toto_signature.IntotoVerifier(fake.FakeVerifier())
-        file_serializer = serialize_by_file.Serializer(
+        file_serializer = file.Serializer(
             self._hasher_factory, allow_symlinks=True
         )
         file_manifest = file_serializer.serialize(sample_model_folder)
@@ -87,7 +87,7 @@ class TestIntotoSignature:
     def test_sign_and_verify_digest_manifest(self, sample_model_folder):
         signer = in_toto_signature.IntotoSigner(fake.FakeSigner())
         verifier = in_toto_signature.IntotoVerifier(fake.FakeVerifier())
-        file_serializer = serialize_by_file.Serializer(
+        file_serializer = file.Serializer(
             self._hasher_factory, allow_symlinks=True
         )
         file_manifest = file_serializer.serialize(sample_model_folder)
